Add 3 new question types to the game.

See merge request cse1105/2021-2022/team-repositories/oopp-group-53/repository-template!62

+++ commons/src/main/java/commons/GameState.java
@@ -23,8 +23,8 @@ public class GameState {
 
     public List<Emote> emotes;
 
-    public long timeToAnswer;//time it took them to answer in milliseconds
-    public long timeOfReceival;//time of receiving the question in milliseconds
+    public double timeToAnswer;//time it took them to answer in milliseconds
+    public double timeOfReceival;//time of receiving the question in milliseconds
 
     public GameState() {
